Have module for fixtures
